Break into smaller files; use browserify
