+++ src/index.js
@@ -1,0 +1,31 @@
+var L = require('leaflet'),
+	Control = require('./control'),
+	Nominatim = require('./geocoders/nominatim'),
+	Bing = require('./geocoders/bing'),
+	MapQuest = require('./geocoders/mapquest'),
+	Mapbox = require('./geocoders/mapbox'),
+	What3Words = require('./geocoders/what3words'),
+	Google = require('./geocoders/google'),
+	Photon = require('./geocoders/photon');
+
+module.exports = L.Util.extend(Control.class, {
+	Nominatim: Nominatim.class,
+	nominatim: Nominatim.factory,
+	Bing: Bing.class,
+	bing: Bing.factory,
+	MapQuest: MapQuest.class,
+	mapQuest: MapQuest.factory,
+	Mapbox: Mapbox.class,
+	mapbox: Mapbox.factory,
+	What3Words: What3Words.class,
+	what3words: What3Words.factory,
+	Google: Google.class,
+	google: Google.factory,
+	Photon: Photon.class,
+	photon: Photon.factory
+});
+
+L.Util.extend(L.Control, {
+	Geocoder: module.exports,
+	geocoder: Control.factory
+});
